Article-07/hello_django/requirements.txt
Django==2.0.5
Pillow==5.1.0
colorific==0.3
django-extra-views==0.11.0
